+++ src/daemon/config/netdata-conf.h
@@ -17,6 +17,4 @@ bool netdata_conf_load(char *filename, char overwrite_used, const char **user);
 #include "netdata-conf-web.h"
 #include "netdata-conf-cloud.h"
 
-#include "daemon/common.h"
-
 #endif //NETDATA_DAEMON_NETDATA_CONF_H
