+++ src/database/engine/pdc.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
-#define NETDATA_RRD_INTERNALS
+
 #include "pdc.h"
 #include "dbengine-compression.h"
