+++ src/database/engine/rrdengine.c
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
-#define NETDATA_RRD_INTERNALS
 
 #include "rrdengine.h"
 #include "pdc.h"
