+++ src/database/rrdfunctions.c
@@ -1,7 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#define NETDATA_RRD_INTERNALS
-
 #include "rrd.h"
 #include "rrdfunctions-internals.h"
