+++ src/database/rrdhost-status.c
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include "rrdhost-status.h"
-#include "stream-receiver-internals.h"
-#include "stream-sender-internals.h"
+#include "streaming/stream-receiver-internals.h"
+#include "streaming/stream-sender-internals.h"
 
 ENUM_STR_MAP_DEFINE(RRDHOST_DB_STATUS) = {
     { RRDHOST_DB_STATUS_INITIALIZING, "initializing" },
