+++ src/plugins.d/pluginsd_parser.c
@@ -2,6 +2,7 @@
 
 #include "pluginsd_internals.h"
 #include "streaming/stream-replication-receiver.h"
+#include "database/rrddim-collection.h"
 
 static inline PARSER_RC pluginsd_set(char **words, size_t num_words, PARSER *parser) {
     int idx = 1;
